feat : 사용자에서 쿠폰 발급, 조회, 사용하는 API 구현 완료

+++ src/main/java/projectbuildup/mivv/domain/worthyConsumption/entity/Condition.java
@@ -21,7 +21,7 @@ public class Condition {
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDate issuableCouponEndDate;
     @NonNull
-    private int lastMonthAmount;
+    private Long lastMonthAmount;
     //@OneToOne(mappedBy = "Condition", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     //private WorthyConsumption worthyConsumption;
     @Setter
